+++ plugin/app/Services/UpdateCharonService.php
@@ -3,10 +3,13 @@
 namespace TTU\Charon\Services;
 
 use Illuminate\Http\Request;
+use TTU\Charon\Events\CharonUpdated;
+use TTU\Charon\Listeners\UpdateCalendarDeadlines;
 use TTU\Charon\Models\Charon;
 use TTU\Charon\Models\Deadline;
 use TTU\Charon\Models\Grademap;
 use TTU\Charon\Repositories\DeadlinesRepository;
+use Zeizig\Moodle\Services\CalendarService;
 use Zeizig\Moodle\Services\GradebookService;
 
 /**
@@ -26,6 +29,8 @@ class UpdateCharonService
     protected $charonGradingService;
     /** @var DeadlinesRepository */
     private $deadlinesRepository;
+    /** @var CalendarService */
+    private $calendarService;
 
     /**
      * UpdateCharonService constructor.
@@ -35,19 +40,22 @@ class UpdateCharonService
      * @param  DeadlineService  $deadlineService
      * @param  DeadlinesRepository  $deadlinesRepository
      * @param  CharonGradingService  $charonGradingService
+     * @param  CalendarService $calendarService
      */
     public function __construct(
         GrademapService $grademapService,
         GradebookService $gradebookService,
         DeadlineService $deadlineService,
         DeadlinesRepository $deadlinesRepository,
-        CharonGradingService $charonGradingService
+        CharonGradingService $charonGradingService,
+        CalendarService $calendarService
     ) {
         $this->grademapService     = $grademapService;
         $this->gradebookService    = $gradebookService;
         $this->deadlineService     = $deadlineService;
         $this->deadlinesRepository = $deadlinesRepository;
         $this->charonGradingService = $charonGradingService;
+        $this->calendarService = $calendarService;
     }
 
     /**
@@ -99,8 +107,10 @@ class UpdateCharonService
     public function updateDeadlines($request, $charon)
     {
         $oldDeadlines = $charon->deadlines;
-        $this->deadlinesRepository->deleteAllDeadlinesForCharon($charon->id);
-        $this->deadlinesRepository->deleteAllCalendarEventsForCharon($charon->id);
+        $charonId = $charon->id;
+
+        $this->deadlinesRepository->deleteAllDeadlinesForCharon($charonId);
+        $this->deadlinesRepository->deleteAllCalendarEventsForCharon($charonId);
 
         // Create new deadlines
         if ($request->deadlines !== null) {
@@ -108,9 +118,10 @@ class UpdateCharonService
                 $this->deadlineService->createDeadline($charon, $deadline);
             }
         }
-
         $charon->load('deadlines');
-
+        $event = new CharonUpdated($charon);
+        $eventAdder = new UpdateCalendarDeadlines($this->calendarService);
+        $eventAdder->handle($event);
         return $this->deadlinesAreNew($oldDeadlines, $charon->deadlines);
     }
 
@@ -161,7 +172,7 @@ class UpdateCharonService
         }
 
         $grademap->name = $newGrademap['grademap_name'];
-        $grademap->persistent = $grademap->grade_type_code > 1000 && isset($newGrademap['persistent']) && (bool) $newGrademap['persistent'];
+        $grademap->persistent = $grademap->grade_type_code > 1000 && isset($newGrademap['persistent']) && $newGrademap['persistent'] === '1';
         $grademap->save();
 
         $oldMax = $grademap->gradeItem->grademax;